apps/backend/pyproject.toml
[tool.poetry]
name = "boilerplate-be"
version = "0.1.0"
description = "Boilerplate"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.7"
graphene-django = "^2.15.0"
django-storages = "^1.11.1"
boto3 = "^1.17.35"
django-graphql-jwt = "^0.3.1"
django-cors-headers = "^3.7.0"
python-dotenv = "^0.15.0"
django-filter = "^2.4.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
Pillow = "^8.1.2"
stringcase = "^1.2.0"

[tool.poetry.dev-dependencies]
ipython = "^7.21.0"
pdbpp = "^0.10.2"
flake8 = "^3.9.0"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
